+++ Source/Internal/Interfaces/Unreal/SDK/WBP_UI_DetectionWidgetContainer_classes.hpp
@@ -17,12 +17,12 @@
 SDK_NAMESPACE_START
 
 // WidgetBlueprintGeneratedClass WBP_UI_DetectionWidgetContainer.WBP_UI_DetectionWidgetContainer_C
-// 0x0010 (0x02F0 - 0x02E0)
+// 0x0010 (0x0350 - 0x0340)
 class UWBP_UI_DetectionWidgetContainer_C final : public UPD3DetectionMeterContainer
 {
 public:
-	struct FPointerToUberGraphFrame               UberGraphFrame;                                    // 0x02E0(0x0008)(ZeroConstructor, Transient, DuplicateTransient)
-	float                                         CollapseDelay;                                     // 0x02E8(0x0004)(Edit, BlueprintVisible, ZeroConstructor, DisableEditOnInstance, IsPlainOldData, NoDestructor, HasGetValueTypeHash)
+	struct FPointerToUberGraphFrame               UberGraphFrame;                                    // 0x0340(0x0008)(ZeroConstructor, Transient, DuplicateTransient)
+	double                                        CollapseDelay;                                     // 0x0348(0x0008)(Edit, BlueprintVisible, ZeroConstructor, DisableEditOnInstance, IsPlainOldData, NoDestructor, HasGetValueTypeHash)
 
 public:
 	void ExecuteUbergraph_WBP_UI_DetectionWidgetContainer(int32 EntryPoint);
